reproducible Debian: let maintenance job cleanup old build service logfiles

+++ bin/reproducible_maintenance.sh
@@ -207,16 +207,17 @@ if [ -d /srv/workspace/pbuilder/ ] ; then
 fi
 
 # delete build services logfiles
-if [ -d /var/lib/jenkins/userContent/reproducible/debian/build_service/ ] ; then
-	echo "$(date -u) - Deleting pbuilder build directories, older than 3 days."
-	OLDSTUFF=$(find /var/lib/jenkins/userContent/reproducible/debian/build_service/ -maxdepth 2 -regex '.*/[0-9]+' -type d -mtime +2 -exec ls -lad {} \; || true)
-	if [ ! -z "$OLDSTUFF" ] ; then
-		echo
-		echo "Old pbuilder build directories found in /var/lib/jenkins/userContent/reproducible/debian/build_service/"
-		echo -n "$OLDSTUFF"
-		find /var/lib/jenkins/userContent/reproducible/debian/build_service/ -maxdepth 2 -regex '.*/[0-9]+' -type d -mtime +2 -exec sudo rm -rf --one-file-system {} \; || true
-		echo
-		DIRTY=true
+if [ "$HOSTNAME" = "$MAINNODE" ] ; then
+	if [ -d /var/lib/jenkins/userContent/reproducible/debian/build_service/ ] ; then
+		echo "$(date -u) - Deleting logfiles from build services directories, older than 3 days."
+		OLDSTUFF=$(find /var/lib/jenkins/userContent/reproducible/debian/build_service/ -maxdepth 2 -regex '.*/[0-9]+' -type d -mtime +2 -exec ls -lad {} \; || true)
+		if [ ! -z "$OLDSTUFF" ] ; then
+			echo
+			echo "Old logfiles cleaned in /var/lib/jenkins/userContent/reproducible/debian/build_service/"
+			echo -n "$OLDSTUFF"
+			find /var/lib/jenkins/userContent/reproducible/debian/build_service/ -maxdepth 2 -regex '.*/[0-9]+' -type d -mtime +2 -exec sudo rm -rf --one-file-system {} \; || true
+			echo
+		fi
 	fi
 fi
